프로젝트 패키지 재구성
