+++ src/main/java/com/bipa/bizsurvey/domain/admin/enums/ClaimType.java
@@ -1,0 +1,17 @@
+package com.bipa.bizsurvey.domain.admin.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum ClaimType {
+    POST("게시글"),
+    COMMENT("댓글"),
+    CHILD_COMMENT("대댓글");
+
+    private final String value;
+
+
+    ClaimType(String value) {
+        this.value = value;
+    }
+}
