--- src/main/java/com/bipa/bizsurvey/domain/answer/enums/Correct.java
@@ -1,20 +1,0 @@
-package com.bipa.bizsurvey.domain.answer.enums;
-
-
-import lombok.Getter;
-
-@Getter
-public enum Correct {
-
-    YES("정답"),
-    NO("오답");
-
-
-    private final String value;
-
-
-    Correct(String value) {
-        this.value = value;
-    }
-
-}
