--- src/main/java/com/bipa/bizsurvey/domain/childComment/domain/ChildComment.java
@@ -1,34 +1,0 @@
-package com.bipa.bizsurvey.domain.childComment.domain;
-
-import com.bipa.bizsurvey.domain.comment.domain.Comment;
-import com.bipa.bizsurvey.domain.user.domain.User;
-import com.bipa.bizsurvey.global.common.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "child_comment")
-public class ChildComment extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "child_comment_id")
-    private Long id;
-
-
-    @Column(nullable = false)
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    Comment comment;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    User user;
-
-}
