--- src/main/java/com/bipa/bizsurvey/domain/claim/enums/ClaimType.java
@@ -1,17 +1,0 @@
-package com.bipa.bizsurvey.domain.claim.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum ClaimType {
-    POST("게시글"),
-    COMMENT("댓글"),
-    CHILD_COMMENT("대댓글");
-
-    private final String value;
-
-
-    ClaimType(String value) {
-        this.value = value;
-    }
-}
