--- src/main/java/com/bipa/bizsurvey/domain/comment/domain/Comment.java
@@ -1,35 +1,0 @@
-package com.bipa.bizsurvey.domain.comment.domain;
-
-import com.bipa.bizsurvey.domain.post.domain.Post;
-import com.bipa.bizsurvey.domain.user.domain.User;
-import com.bipa.bizsurvey.global.common.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "comment")
-public class Comment extends BaseEntity {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id")
-    private Long id;
-
-    @Column(nullable = false)
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-}
