+++ src/main/java/com/bipa/bizsurvey/domain/community/api/PostController.java
@@ -1,0 +1,26 @@
+package com.bipa.bizsurvey.domain.community.api;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/community")
+public class PostController {
+
+    // CREATE
+//    @PostMapping("/createPost")
+//    public ResponseEntity<?> createPost(@Valid @RequestBody CreatePostRequestt createPostRequest,
+//                                        @AuthenticationPrincipal Long userId
+//                                        ){
+//
+//
+//
+//    }
+
+    //
+
+
+}
