+++ src/main/java/com/bipa/bizsurvey/domain/community/domain/VoteAnswer.java
@@ -1,0 +1,28 @@
+package com.bipa.bizsurvey.domain.community.domain;
+
+import com.bipa.bizsurvey.domain.community.domain.Vote;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Table(name = "vote_answer")
+public class VoteAnswer {
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "vote_answer_id")
+    private Long id;
+
+    @Column(nullable = false)
+    private String answer;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "vote_id")
+    Vote vote;
+}
