+++ src/main/java/com/bipa/bizsurvey/domain/community/repository/PostRepository.java
@@ -1,0 +1,7 @@
+package com.bipa.bizsurvey.domain.community.repository;
+
+import com.bipa.bizsurvey.domain.community.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
