--- src/main/java/com/bipa/bizsurvey/domain/post/domain/Post.java
@@ -1,38 +1,0 @@
-package com.bipa.bizsurvey.domain.post.domain;
-
-import com.bipa.bizsurvey.domain.post.enums.PostType;
-import com.bipa.bizsurvey.domain.user.domain.User;
-import com.bipa.bizsurvey.global.common.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.ColumnDefault;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "post")
-public class Post extends BaseEntity {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_id")
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @ColumnDefault("0")
-    private String count;
-
-    @Enumerated(EnumType.STRING)
-    private PostType type;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-}
