--- src/main/java/com/bipa/bizsurvey/domain/postImages/domain/PostImages.java
@@ -1,28 +1,0 @@
-package com.bipa.bizsurvey.domain.postImages.domain;
-
-import com.bipa.bizsurvey.domain.post.domain.Post;
-import com.bipa.bizsurvey.global.common.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "post_images")
-public class PostImages extends BaseEntity {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "postimages_id")
-    private Long id;
-
-    private String imgName;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-}
