--- src/main/java/com/bipa/bizsurvey/domain/surveyPost/domain/SurveyPost.java
@@ -1,43 +1,0 @@
-package com.bipa.bizsurvey.domain.surveyPost.domain;
-
-import com.bipa.bizsurvey.domain.post.domain.Post;
-import com.bipa.bizsurvey.domain.survey.domain.Survey;
-import com.bipa.bizsurvey.global.common.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "survey_post")
-public class SurveyPost extends BaseEntity {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "survey_post_id")
-    private Long id;
-
-    @Column(nullable = false)
-    private String deadLine;
-
-    @Column(nullable = false)
-    private int memberCount;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "survey_id")
-    private Survey survey;
-
-
-
-
-
-
-}
