--- src/main/java/com/bipa/bizsurvey/domain/vote/domain/Vote.java
@@ -1,30 +1,0 @@
-package com.bipa.bizsurvey.domain.vote.domain;
-
-import com.bipa.bizsurvey.domain.post.domain.Post;
-import com.bipa.bizsurvey.global.common.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "vote")
-public class Vote extends BaseEntity {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "vote_id")
-    private Long id;
-
-    @Column(nullable = false)
-    private String voteQuestion;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    Post post;
-
-}
