+++ src/main/java/com/bipa/bizsurvey/domain/workspace/domain/SharedSurvey.java
@@ -1,0 +1,35 @@
+package com.bipa.bizsurvey.domain.workspace.domain;
+
+
+import com.bipa.bizsurvey.domain.survey.domain.Survey;
+import com.bipa.bizsurvey.global.common.BaseEntity;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Table(name = "shared_survey")
+public class SharedSurvey extends BaseEntity {
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "shared_survey_id")
+    private Long id;
+
+    private String title;
+
+    @Column(nullable = false)
+    private String deadline;
+
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "survey_id")
+    private Survey survey;
+
+
+}
